+++ CHANGELOG.md
@@ -7,3 +7,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add AdvLayNav select to attribute frontend options.
 - Add Range Slider as option.
 - Render Range Slider filter in frontend with jQuery UI range slider.
+- Ajax layered navigation